+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Fixed MOP not working properly, if player is using analog controller
 - Freshly spawned firewood logs should despawn correctly
 - Fixed "mop generate-list" and "mop load-rules" commands not working as intended
+- Possible bug fix for player dying in Satsuma from low speed impacts
 
 ### Removed